[animation] move AnimationLiveProperties into animation-animanion module.
