+++ lib/js/components/animation-animanion.mjs
@@ -8,6 +8,10 @@ import {
   , CoherenceFunction
 } from '../metamodel.mjs';
 
+import {
+    _BaseComponent
+} from './basics.mjs';
+
 function* enumerate(iterable) {
     let i=0;
     for(const value of iterable) {
@@ -503,5 +507,94 @@ export class Animanion {
         // maybe we got to inject a function into these!
         const momentT = this.keyMomentsKeyToT.get(keyMomentsKey);
         return this.getPropertyValuesMapForAbsoluteT(momentT);
+    }
+}
+
+/**
+ * This is a central class to propagate animation property values
+ * without going via the model. I.e. stuff that is not part of the
+ * model, but derrived from it.
+ * It should be defined before all of its users, so its values
+ * are updated when the users update and read them.
+ * Its users in turn should listen to updates of all of its dependencies,
+ * so they in turn don't miss a value change.
+ * FIXME: This method could (should?) be formalized, so we never miss
+ * changes to dependency values and hence likely make the code overall
+ * leaner. But with `parentAPI.getWidgetById` we can start to explore
+ * this kind of value propagation, without having to create a more generic
+ * mechanism right away.
+ * In general, for each component, no matter if it renders anything, it
+ * should be possible to expose this kind of behavior.
+ */
+export class AnimationLiveProperties extends _BaseComponent {
+    constructor(parentAPI, initAnimanion) {
+        super(parentAPI);
+        this._initAnimanion = initAnimanion;
+        this.animanion = null;
+        this.t = null;
+        this.propertyValuesMap = null;
+    }
+
+    /**
+     * return an instance of Animanion
+     * e.g:
+     *      const outerAnimanion =  new Animanion(null, outerPropertiesGenerator, keyMoments, isLoop);
+     *      return new Animanion(outerAnimanion, innerPropertiesGenerator, keyMoments, isLoop);
+     *
+     * (outerAnimanion, propertiesGenerator, keyMoments, isLoop);
+     */
+    _initAnimanion(keyMoments, isLoop) {
+        throw new Error(`NOT IMPLEMENTED _initAnimanion(${keyMoments}, ${isLoop})`);
+    }
+
+    update(changedMap) {
+        if(changedMap.has('keyMoments') || changedMap.has('isLoop')) {
+            const keyMoments = changedMap.has('keyMoments')
+                    ? changedMap.get('keyMoments')
+                    : this.getEntry('keyMoments')
+              , isLoop = (changedMap.has('isLoop')
+                    ? changedMap.get('isLoop')
+                    : this.getEntry('isLoop')).value
+              ;
+            // when either keyMoments or isLoop change
+            this._initAnimanion(keyMoments, isLoop);
+            this.animanion = this._initAnimanion(keyMoments, isLoop);
+        }
+
+        if(changedMap.has('t') || changedMap.has('keyMoments') || changedMap.has('isLoop')) {
+            const t = (changedMap.has('t')
+                            ? changedMap.get('t')
+                            : this.getEntry('t')).value
+              , propertyValuesMap = this._getPropertyValuesMapForT(t /* between 0 and 1*/)
+              ;
+            this.t = t;
+            this.propertyValuesMap = propertyValuesMap;
+        }
+    }
+
+    get fullDuration() {
+        return this.animanion.fullDuration;
+    }
+
+    get tToKeyMoments() {
+        return this.animanion.tToKeyMoments;
+    }
+
+    _getPropertyValuesMapForT(t) {
+        return this.animanion.getPropertyValuesMapForT(t);
+    }
+
+    getPropertyValuesMapForT(t) {
+        // Not cached, only for current t. t changes potentially very often
+        // and also potentially doesn't repeat much due to floating point precision.
+        if(this.t === t)
+            return this.propertyValuesMap;
+        return this._getPropertyValuesMapForT(t);
+    }
+
+    getPropertyValuesMapForKeyMoment(keyMomentsKey) {
+        // cached until this.animanion is reinitialized, we need these
+        // values repeatedly for ui-elements.
+        return this.animanion.getPropertyValuesMapForKeyMoment(keyMomentsKey);
     }
 }
